Optimize html display
